requirements.txt
Django==1.8
djangorestframework==3.6.0
Markdown==3.1
Pygments==2.3.1
psycopg2==2.8.1
django-filter==1.1.0
